Created NavBar page fro navbar export and notification page.
